backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
openai==0.27.8
openapi-schema-pydantic==1.2.4
langchain==0.0.268
langchain-experimental==0.0.10
# Flask==2.3.2
typing-inspect==0.9.0
typing_extensions==4.5.0
pydantic==1.10.9
pyodbc==4.0.39
tiktoken
# sentence_transformers
chromadb==0.3.29
pandas==2.0.3
# unstructured
faiss-cpu
# faiss-gpu
uuid
datetime